feat: translate md to solid
