+++ src/app.tsx
@@ -6,7 +6,11 @@ import { MetaProvider } from '@solidjs/meta';
 import '@fontsource-variable/montserrat';
 import '@fontsource-variable/saira';
 import './main.css';
+
 import { Banner } from './components/ui/banner';
+import { proseContainerCss } from './styles/prose';
+import { Box } from 'styled-system/jsx';
+import { css } from 'styled-system/css';
 
 /**
  * The root component for the app.
@@ -16,7 +20,13 @@ export default function App() {
     <MetaProvider>
       <Banner>Launching Beta in 2025</Banner>
 
-      <Router root={(props) => <Suspense>{props.children}</Suspense>}>
+      <Router
+        root={(props) => (
+          <Suspense>
+            <Box class={css(proseContainerCss)}>{props.children}</Box>
+          </Suspense>
+        )}
+      >
         <FileRoutes />
       </Router>
     </MetaProvider>
