+++ src/routes/(static)/legal/acknowledgements.md
@@ -1,3 +1,5 @@
+:::main{.prose role=main}
+
 # Acknowledgements
 
 This project was developed using the following open-source libraries:
@@ -6,3 +8,5 @@ This project was developed using the following open-source libraries:
 - [SolidJS](https://www.solidjs.com/)
 - [Rust](https://www.rust-lang.org/)
 - [Panda CSS](https://panda-css.com/)
+
+:::
