+++ src/routes/(static)/legal/copyright-policy.md
@@ -1,3 +1,5 @@
+:::main{.prose role=main}
+
 # Nurl Copyright Policy
 
 ## Effective: April 27, 2024
@@ -7,3 +9,5 @@
 Nurl respects the intellectual property rights of others and expects its users to do the same. It is Nurl's policy, in appropriate circumstances and at its discretion, to disable and/or terminate the accounts of users who repeatedly infringe or are repeatedly charged with infringing the copyrights or other intellectual property rights of others.
 
 All designs, text, graphics, pictures, information, data, software, sound files, other files and the selection and arrangement thereof (collectively, the "Site Materials") are the proprietary property of Nurl or its licensors or users and are protected by U.S. and international copyright laws.
+
+:::
