+++ src/routes/(static)/legal/data-reporting-metrics.md
@@ -1,3 +1,5 @@
+:::main{.prose role=main}
+
 # Nurl Data Request Reporting Metrics
 
 ## Effective: April 27, 2024
@@ -5,3 +7,5 @@
 ## Last Updated: April 27, 2024
 
 As of April 27, 2024, we have not received any data requests for Nurl data.
+
+:::
