+++ src/routes/(static)/legal/group-subscription-policy.md
@@ -1,3 +1,7 @@
+:::main{.prose role=main}
+
 # Group Subscription Policy
 
 This policy is used to group subscription policies together. It is not used to enforce any restrictions.
+
+:::
