+++ src/routes/(static)/legal/how-long-we-keep-your-data.md
@@ -1,3 +1,5 @@
+:::main{.prose role=main}
+
 # How long Nurl keeps your information
 
 We collect certain information as you use Nurl. Our [Privacy Policy](/legal/privacy/) describes the information we collect, how we use and share it, and how you can control your information. This page describes how long we hold on to different types of information and why.
@@ -56,3 +58,5 @@ As provided in our [Privacy Policy](/legal/privacy/), all our users have control
 <br />
 
 If you have any questions about data retention or our deletion practices, or if you wish to exercise control over your information, please reach out to [privacy@nurl.app](mailto:privacy@nurl.app).
+
+:::
