+++ src/routes/(static)/legal/licenses.md
@@ -1,5 +1,9 @@
+:::main{.prose role=main}
+
 # Licenses
 
 ## We :heart: Open Source Software
 
 Licenses of the software we use are coming soon.
+
+:::
