+++ src/routes/(static)/legal/monitization-terms.md
@@ -1,3 +1,5 @@
+:::main{.prose role=main}
+
 # Nurl Monitization Terms
 
 ## Effective: April 27, 2024
@@ -5,3 +7,5 @@
 ## Last Updated: April 27, 2024
 
 Nurl terms will be updated from time to time. Please check back regularly for updates.
+
+:::
