+++ src/routes/(static)/legal/off-platform-behavior.md
@@ -1,5 +1,9 @@
+:::main{.prose role=main}
+
 # Off Platform Behavior
 
 ## Overview
 
 Off platform behavior is a way to track user behavior that happens outside of the platform. This can be useful for tracking user behavior that happens in the real world, or on other platforms.
+
+:::
