+++ src/routes/(static)/legal/other.md
@@ -1,3 +1,7 @@
+:::main{.prose role=main}
+
 # Additional Terms and Policies
 
 We currently have no other policies.
+
+:::
