+++ src/routes/(static)/legal/paid-services.md
@@ -1,3 +1,5 @@
+:::main{.prose role=main}
+
 # Paid services
 
 <!-- Write a list of the following paid services: Netlify, Resend, Github -->
@@ -5,3 +7,5 @@
 - [Netlify](https://www.netlify.com/)
 - [Resend](https://resend.io/)
 - [Github](https://github.com)
+
+:::
