+++ src/routes/[...404].tsx
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  return <div>Not Found</div>;
+}
